# book-api/src/main/java/dev/ducku/bookapi/feedback/FeedbackRequest.java
package dev.ducku.bookapi.feedback;

import jakarta.validation.constraints.*;

public record FeedbackRequest(
        @Positive(message = "200")
        @Min(message = "201", value = 0L)
        @Max(message = "202", value = 5L)
        Double note,
        @NotNull(message = "203")
        @NotEmpty(message = "203")
        @NotBlank(message = "203")
        String comment,
        @NotNull(message = "204")
        Integer bookId) {
}