# book-ui/src/app/services/models/validation-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ValidationRequest {
  email: string;
  firstname: string;
  lastname: string;
  password: string;
}